requirements.txt
toml
python-telegram-bot
pluginbase
pyyaml
praw
requests